backend/pyproject.toml
[tool.poetry]
name = "bridgebloc-backend"
version = "0.0.1"
description = ""
authors = [
  "Isaac Adewumi <ienioladewumi@gmail.com>",
  "Tobiloba Emmanuel",
  "Olayinka Ganiyu"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-cors-headers = "^4.2.0"
django = "^4.2.4"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
web3 = "^6.8.0"
huey = "^2.4.5"
psycopg2-binary = "^2.9.7"
django-environ = "^0.10.0"
django-extra-checks = "^0.13.3"
django-stubs-ext = "^4.2.2"
redis = "^4.6.0"
whitenoise = "^6.5.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.282"
mypy = "^1.4.1"
black = "^23.7.0"
isort = "^5.12.0"
pylint = "^2.17.5"
django-stubs = "^4.2.3"
djangorestframework-stubs = "^3.14.2"
pytest = "^7.4.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
ipdb = "^0.13.13"
pytest-deadfixtures = "^2.2.1"
pytest-sugar = "^0.9.7"
pylint-django = "^2.5.3"
pyinstrument = "^4.5.1"
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
skip-string-normalization = true
target-version = ['py39']
exclude = '''
  /(
    | \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
'''

[tool.ruff]
line-length = 119
select = [
    "E",
    "F",
    "W",
    "C",
    "N",
    "B",
    "G",
    "ISC",
    "ICN",
    "T",
    "UP",
    "YTT",
    "S",
    "BLE",
    "FBT",
    "A",
    "Q",
    "ERA",
    "PGH",
    "RET",
    "SIM",
    "TID",
    "ARG",
    "PL",
    "PIE",
    "PYI",
    "COM",
    "DTZ",
    "EXE",
    "INP",
    "PT",
    "TCH",
    "PTH",
    "INT",
    "TRY",
    "RSE",
    "SLF",
    "RUF",
]
exclude = ["migrations", ".venv"]
ignore = ["TRY003", "G004", "TRY301"]

[tool.ruff.per-file-ignores]
"**/test_**.py" = ["S101"]
"bridgebloc/conf/settings.py" = ["E402"]

[tool.ruff.flake8-quotes]
inline-quotes = "single"
docstring-quotes = "double"
multiline-quotes = "double"

[tool.isort]
profile = "black"
line_length = 119
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
known_django = "django"
combine_as_imports = true
lines_between_sections = 1
include_trailing_comma = true
known_first_party = "bridgebloc"
known_rest_framework = "rest_framework"
skip_glob = ["**/.venv/*", "**/migrations/*"]
sections = ["FUTURE","STDLIB","THIRDPARTY","DJANGO", "REST_FRAMEWORK", "FIRSTPARTY","LOCALFOLDER"]

[tool.mypy]
python_version = "3.11"
exclude = "^.venv/"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

ignore_missing_imports = true
check_untyped_defs = true
no_implicit_reexport = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_untyped_calls = true
warn_unreachable = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[tool.django-stubs]
django_settings_module = "bridgebloc.conf.settings"

[[tool.mypy.overrides]]
module = [
  "environ",
  "factory.*",
]
ignore_missing_imports = true

# -- Ignore typing for tests --
[[tool.mypy.overrides]]
module = []
ignore_errors = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bridgebloc.conf.settings"
python_files = ["test_*.py"]

[tool.coverage.run]
branch = true
source = ["bridgebloc"]
omit = ["*/tests/*", "*/migrations/*", "*/conf/*", "conftest.py"]